+++ Public/Get-LMUnmonitoredDevice.ps1
@@ -1,4 +1,4 @@
-Function Get-LMUnmonitoredDevices {
+Function Get-LMUnmonitoredDevice {
 
     [CmdletBinding()]
     Param (
@@ -41,7 +41,7 @@ Function Get-LMUnmonitoredDevices {
                 #Stop looping if single device, no need to continue
                 If ($PSCmdlet.ParameterSetName -eq "Id") {
                     $Done = $true
-                    Return $Response
+                    Return (Add-ObjectTypeInfo -InputObject $Response -TypeName "LogicMonitor.UnmonitoredDevice" )
                 }
                 #Check result size and if needed loop again
                 Else {
@@ -60,7 +60,7 @@ Function Get-LMUnmonitoredDevices {
                 }
             }
         }
-        Return $Results
+        Return (Add-ObjectTypeInfo -InputObject $Results -TypeName "LogicMonitor.UnmonitoredDevice" )
     }
     Else {
         Write-Error "Please ensure you are logged in before running any commands, use Connect-LMAccount to login and try again."