+++ Public/Set-LMUser.ps1
@@ -145,9 +145,8 @@ Function Set-LMUser {
                     adminGroupIds       = $AdminGroupIds
 
                 }
-
                 #Remove empty keys so we dont overwrite them
-                @($Data.keys) | ForEach-Object { if (-not $Data[$_]) { $Data.Remove($_) } }
+                @($Data.keys) | ForEach-Object { if ([string]::IsNullOrEmpty($Data[$_])) { $Data.Remove($_) } }
 
                 $Data = ($Data | ConvertTo-Json)